+++ deepinv/loss/metric/perceptual.py
@@ -11,7 +11,9 @@ class LPIPS(Metric):
     Calculates the LPIPS :math:`\text{LPIPS}(\hat{x},x)` where :math:`\hat{x}=\inverse{y}`.
 
     Computes the perceptual similarity between two images, based on a pre-trained deep neural network.
-    Uses implementation from `pyiqa <https://pypi.org/project/pyiqa/>`_.
+    Uses implementation from `torchmetrics <https://lightning.ai/docs/torchmetrics/stable/image/learned_perceptual_image_patch_similarity.html>`_.
+
+    The inputs `x_net`, `x` must both have 3 channels and be in `[0, 1]`.
 
     .. note::
 
@@ -21,8 +23,8 @@ class LPIPS(Metric):
 
     >>> from deepinv.utils import load_example
     >>> from deepinv.loss.metric import LPIPS
-    >>> m = LPIPS() # doctest: +IGNORE_RESULT
-    >>> x = load_example("celeba_example.jpg", img_size=128)
+    >>> m = LPIPS()
+    >>> x = torch.ones(2, 3, 32, 32)
     >>> x_net = x - 0.01
     >>> m(x_net, x) # doctest: +ELLIPSIS
     tensor([...])
@@ -33,23 +35,34 @@ class LPIPS(Metric):
     :param str reduction: a method to reduce metric score over individual batch scores. ``mean``: takes the mean, ``sum`` takes the sum, ``none`` or None no reduction will be applied (default).
     :param str norm_inputs: normalize images before passing to metric. ``l2`` normalizes by :math:`\ell_2` spatial norm, ``min_max`` normalizes by min and max of each input.
     :param bool check_input_range: if True, ``pyiqa`` will raise error if inputs aren't in the appropriate range ``[0, 1]``.
-    :param bool as_loss: if True, returns LPIPS as a loss. Default: False.
     :param int, tuple[int], None center_crop: If not `None` (default), center crop the tensor(s) before computing the metrics.
         If an `int` is provided, the cropping is applied equally on all spatial dimensions (by default, all dimensions except the first two).
         If `tuple` of `int`, cropping is performed over the last `len(center_crop)` dimensions. If positive values are provided, a standard center crop is applied.
         If negative (or zero) values are passed, cropping will be done by removing `center_crop` pixels from the borders (useful when tensors vary in size across the dataset).
     """
 
-    def __init__(self, device="cpu", check_input_range=False, as_loss=False, **kwargs):
+    def __init__(self, device="cpu", **kwargs):
         super().__init__(**kwargs)
-        pyiqa = import_pyiqa()
-        self.lpips = pyiqa.create_metric(
-            "lpips", check_input_range=check_input_range, device=device, as_loss=as_loss
+        from torchmetrics.image.lpip import LearnedPerceptualImagePatchSimilarity
+
+        self.lpips = LearnedPerceptualImagePatchSimilarity(
+            net_type="alex", normalize=True, reduction="none"
         ).to(device)
-        self.lower_better = self.lpips.lower_better
+        self.lower_better = True
 
     def metric(self, x_net, x, *args, **kwargs):
-        return self.lpips(x_net, x).squeeze(-1)
+        if not (
+            torch.all((0.0 <= x_net) & (x_net <= 1.0))
+            and torch.all((0.0 <= x_net) & (x_net <= 1.0))
+        ):
+            raise ValueError("LPIPS metric requires x_net and x to be between 0 and 1.")
+
+        if not (x_net.shape[1] == x.shape[1] == 3):
+            raise ValueError(
+                f"LPIPS metric only supports 3-channel input, but got channels for x_net, x as {x_net.shape[1]}, {x.shape[1]}."
+            )
+
+        return self.lpips(x_net, x)
 
 
 class NIQE(Metric):